Improved Riija Sword durability healing
Players have the ability to 'look' at a Riija Sword and mend it one
point of durability thanks to a bug that turned out to be rather cool
and lore appropriate. Instead of removing that feature, this pull gets
rid of the 'grindy' portion of mending and just fully mends the sword if
you cast a Riija spell or look at it. Instead of clicking hundreds of
times, you'll only have to click once.

This shouldn't negatively affect the durability system for Riija swords,
because they weren't really breaking due to durability to begin with
(almost all of them are just held for spellpower purposes), so this is
just a cool little feature.