Shield & Damage rod improvements
Shield and Damage rods are now SpellRods, and drain when in the inn so
that only builders and those with prior setup have access to charges.
This is an attempt to keep them out of character train use. We'll see
how it goes.

Shield and damage rods get their own icons.

Shield rod bounds damage to 10 max instead of 1 now.