Reincarnation Training Points total function
So after about a year of observation (and before that a bunch of fine
tuning), we've learned a lot about the reincarnation system. Mainly,
that people just want to 'get it done' and do the first nine reincs as
Demon to just skip all of it.

I think we've come up with a good change to fix that.

The main core of the idea is to give you, upon reincarnation, the entire
sum of your training points back. This is calculated by the meditation
ratio multiplied by the percent in all your skills and spells. A new
character might expect 6k, a half-built 50k, a decent toon 100k or more,
and a fully built character ranks in at somewhere around 200k. Of
course, this depends on the character you made.

We could also possibly give you a 5% or 10% bonus so that reincarnating
actually increases your total TP pool.

So what would reincarnations be for if you 'reroll' to the same or
slightly more training points? That's the big question. The first and
most obvious is that it allows you to change your race. You can reroll
and retain the 'wisdom' of your lives.

But the rest is unclear. What we're envisioning right now is that
reincarnations will still start you over at a lesser HP than you have.
That makes sense with a whole 'new life' angle. However, we'd like to
take away the need to do this ten times over.

You know how you get 3x Training Points when you kill tough monsters
while you're at Max HP? The idea I'd like to run by you guys is that
reincarnation becomes an optional thing that gives you a 5% or 10% TP
bonus on your entire existing pool. Meanwhile, the extra stat points you
used to get are instead purchased with increasing sums of TP. For
example, your first extra stat point might cost 10,000 TP. The next
20,000. that kind of thing.

This means a currently built 'r0' character could fight monsters, do
quests, and explore new content until they generated the TP to buy those
stat points. No reincarnation required. This would fit well with areas
like the desert and the maze, where often fully built characters are
taking on huge dungeons from which they get no 'experience' since
they're at max hp. So instead of wasting all that, you're instead buying
extra stat points.

And so the 'reincarnation' process becomes not only a way to try out new
races and builds, but to get that 5% or 10% bonus on your pool. Then you
spend that TP you just gained by reincarnating to buy the extra stat
point. Or don't spend it. It'll be up to you.

What do you guys think?